Events Code Separation
