+++ services/function/events_read.go
@@ -1,0 +1,14 @@
+package function
+
+//S3ReadEvent S3 trigger of a lambda function
+type S3ReadEvent struct {
+	Bucket      *string
+	StatementId *string
+}
+
+//HTTPReadEvent  HTTP (API Gateway) trigger of a lambda function
+type HTTPReadEvent struct {
+	ApiId      *string
+	ResourceId *string
+	Method     *string
+}
